Fix til.py broke on macOS; re-organize the directory
The error was not with MacOS but with the default editor (vim) used to
edit the files. In a vim instance where swap files are allowed, previous
method of editing failed. The revamped code now works irrespective of
how the vim is set up.

Further, organized the content into more sensible directories.

Signed-off-by: Vivek Rai <mail@raivivek.in>